Move getCircleCiShardTestFiles to utils.js
