+++ build-system/pr-check/utils.js
@@ -34,6 +34,7 @@ const {
 const {cyan, green, yellow} = require('../common/colors');
 const {exec, execOrDie, execOrThrow, execWithError} = require('../common/exec');
 const {getLoggingPrefix, logWithoutTimestamp} = require('../common/logging');
+const {getStdout} = require('../common/process');
 const {replaceUrls} = require('../tasks/pr-deploy-bot-utils');
 
 const UNMINIFIED_CONTAINER_DIRECTORY = 'unminified';
@@ -306,6 +307,21 @@ async function processAndStoreBuildToArtifacts() {
   execOrDie(`du -sh ${ARTIFACT_FILE_NAME}`);
 }
 
+/**
+ * Returns list of test files that CircleCI should execute in a parallelized job shard.
+ *
+ * @param {!Array<string>} globs array of glob strings for finding test file paths.
+ * @return {string} a comma separated list of test file paths.
+ */
+function getCircleCiShardTestFiles(globs) {
+  const joinedGlobs = globs.map((glob) => `"${glob}"`).join(' ');
+  return getStdout(
+    `circleci tests glob ${joinedGlobs} | circleci tests split --split-by=timings`
+  )
+    .trim()
+    .replace(/\s+/g, ',');
+}
+
 module.exports = {
   abortTimedJob,
   printChangeSummary,
@@ -321,4 +337,5 @@ module.exports = {
   storeModuleBuildToWorkspace,
   storeExperimentBuildToWorkspace,
   processAndStoreBuildToArtifacts,
+  getCircleCiShardTestFiles,
 };
